+++ lib/useful_widgets.dart
@@ -1,4 +1,5 @@
 library useful_widgets;
 
 export 'widgets/future/future_widget.dart';
+export 'widgets/refresh/refresh_widget.dart';
 export 'widgets/search/search_widget.dart';