+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.0 - 2019-01-03
 
-* Added `FutureWidget`, `SearchWidget`
+* Added `FutureWidget`, `SearchWidget`, `RefreshWidget`